+++ app/routes/components/Button/Demo.tsx
@@ -1,5 +1,11 @@
 import { Button } from "src/components/Button";
 
 export function Demo() {
-  return <Button>Button</Button>;
+  return (
+    <div style={{ display: 'flex', gap: '1rem', alignItems: 'center' }}>
+      <Button size="sm">Oh</Button>
+      <Button>Hello</Button>
+      <Button size="lg" variant="filled">World</Button>
+    </div>
+  );
 }