+++ app/routes/components/Button/Demo_source_code.codegen.ts
@@ -2,6 +2,12 @@ export const demoSourceCode = ''
   + 'import { Button } from "src/components/Button";\n'
   + '\n'
   + 'export function Demo() {\n'
-  + '  return <Button>Button</Button>;\n'
+  + '  return (\n'
+  + "    <div style={{ display: 'flex', gap: '1rem', alignItems: 'center' }}>\n"
+  + '      <Button size="sm">Oh</Button>\n'
+  + '      <Button>Hello</Button>\n'
+  + '      <Button size="lg" variant="filled">World</Button>\n'
+  + '    </div>\n'
+  + '  );\n'
   + '}\n'
   ;