+++ app/routes/components/Button/changelog.codegen.ts
@@ -1,6 +1,8 @@
 export const changelog = ''
   + '- 2025-09-04: Initial version.\n'
-  + '- 2025-09-05: Use ThemeProvider for theming, support the props `leftSection` and `rightSection`.\n'
+  + '- 2025-09-05: Use ThemeProvider for theming, support the props `leftSection` and\n'
+  + '  `rightSection`.\n'
   + '- 2025-09-08: Support the prop `ref`.\n'
   + '- 2025-09-09: Support the prop `variant`.\n'
+  + '- 2025-09-18: Support the prop `size` (`sm`, `md` and `lg`).\n'
   ;