+++ src/components/Button/Button.tsx
@@ -8,6 +8,7 @@ import styles from './Button.module.scss';
 
 export type ButtonProps = ComponentProps<'button'> & {
   variant?: 'default' | 'filled';
+  size?: 'sm' | 'md' | 'lg';
   leftSection?: React.ReactNode;
   rightSection?: React.ReactNode;
 };
@@ -22,6 +23,7 @@ export function Button(props: ButtonProps) {
     leftSection,
     rightSection,
     variant = 'default',
+    size = 'md',
     ...rest
   } = props;
 
@@ -42,9 +44,25 @@ export function Button(props: ButtonProps) {
       '--button-border-color': theme.colors.blue['700'],
       '--button-bg-hover': theme.colors.blue['700'],
     } as CSSProperties,
-    {
+    size === 'sm' && {
+      '--button-height': '2rem',
+      '--button-padding-x': '0.75rem',
+      '--button-section-gap': '0.5rem',
+      '--button-font-size': '0.875rem',
+    } as CSSProperties,
+    size === 'md' && {
       '--button-height': '2.25rem',
       '--button-padding-x': '1rem',
+      '--button-section-gap': '0.625rem',
+      '--button-font-size': '1rem',
+    } as CSSProperties,
+    size === 'lg' && {
+      '--button-height': '2.5rem',
+      '--button-padding-x': '1.25rem',
+      '--button-section-gap': '0.75rem',
+      '--button-font-size': '1.125rem',
+    } as CSSProperties,
+    {
       '--button-radius': '0.375rem',
     } as CSSProperties,
     style,