+++ src/components/Button/CHANGELOG.md
@@ -1,4 +1,6 @@
 - 2025-09-04: Initial version.
-- 2025-09-05: Use ThemeProvider for theming, support the props `leftSection` and `rightSection`.
+- 2025-09-05: Use ThemeProvider for theming, support the props `leftSection` and
+  `rightSection`.
 - 2025-09-08: Support the prop `ref`.
 - 2025-09-09: Support the prop `variant`.
+- 2025-09-18: Support the prop `size` (`sm`, `md` and `lg`).